fix: SignatureListの型を修正

+++ model/home.go
@@ -42,9 +42,9 @@ type Vocabulary struct {
 }
 
 type SignatureList struct {
-	ID        uint      `gorm:"primaryKey;autoIncrement" json:"id"`
+	ID        uint64      `gorm:"primaryKey;autoIncrement" json:"id"`
 	Content   string    `gorm:"type:text;not null" json:"content"`
-	UserID    uint      `gorm:"not null" json:"user_id"`
+	UserID    uint64      `gorm:"not null" json:"user_id"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
 }
